update: optimize sth;

+++ deploy/update.py
@@ -6,6 +6,7 @@ import os
 import shutil
 import stat
 import pathlib
+import traceback
 import zipfile
 
 import httpx
@@ -32,6 +33,11 @@ class GitHandler:
 
     # src_url = f"{self.speedup_prefix}https://github.com/{self.github_author}/{proj_name}/archive/refs/heads/{branch}.zip"
 
+    def get_version_info(self, ver):
+        resp = self.sess.get(f"{self.commit_api}/{ver}", headers=headers)
+        resp_json = resp.json()
+        return resp_json
+
     def check_changed_files(self, ver):
         print(Fore.BLUE + "[ 检查版本中.. ]")
         resp = self.sess.get(self.commit_api, headers=headers)
@@ -43,18 +49,18 @@ class GitHandler:
         ver_index = vers.index(ver)
         valid_vers = vers[:ver_index]
         files = []
-        for _ver in tqdm(valid_vers, total=len(valid_vers), ncols=80, ascii=True,
-                         desc=Fore.BLUE + "[ 检查需要更新的代码中.. ]"):
-            resp = self.sess.get(f"{self.commit_api}/{_ver}", headers=headers)
-            resp_json = resp.json()
+        print(Fore.BLUE + f"[ 检查需要更新的代码中.. ]")
+        for _ver in valid_vers:
+            resp_json = self.get_version_info(_ver)
             files.extend(list(map(lambda _: _["filename"], resp_json["files"])))
+            print(Fore.GREEN + f"[ {_ver[:8]} ] {resp_json['commit']['message']}")
         return vers[0], list(set(files))
 
-    def download_src_code(self):
+    def download_src_code(self, _url=None, zip_name="src.zip"):
         """proj less than 1Mb, actually just take little second"""
         temp_p.mkdir(exist_ok=True)
-        zip_file = temp_p.joinpath("src.zip")
-        with self.sess.stream("GET", self.src_url, follow_redirects=True) as resp:
+        zip_file = temp_p.joinpath(zip_name)
+        with self.sess.stream("GET", _url or self.src_url, follow_redirects=True) as resp:
             with open(zip_file, 'wb') as f:
                 size = 50 * 1024
                 for chunk in tqdm(resp.iter_bytes(size), ncols=80, ascii=True, desc=Fore.BLUE + "[ 下载代码文件中.. ]"):
@@ -70,6 +76,7 @@ class Proj:
     git_handler = GitHandler(github_author, name, branch)
     ver = ""
     first_flag = False
+    local_ver = None
     changed_files = []
 
     def check_existed_version(self):
@@ -81,7 +88,7 @@ class Proj:
                 return f.read().strip()
 
     def check(self):
-        local_ver = self.check_existed_version()
+        self.local_ver = local_ver = self.check_existed_version()
         self.ver, self.changed_files = self.git_handler.check_changed_files(local_ver)
 
     def local_update(self):
@@ -92,7 +99,9 @@ class Proj:
         def move(src, dst):
             if src.is_dir() and dst.exists():
                 shutil.rmtree(dst, ignore_errors=True)
-                os.rmdir(dst)
+                if dst.exists():  # fix
+                    os.rmdir(dst)
+            dst.parent.mkdir(exist_ok=True)
             shutil.move(src, dst)
 
         if not self.first_flag and not self.changed_files:
@@ -116,12 +125,31 @@ class Proj:
                                      ncols=80, ascii=True, desc=Fore.BLUE + "[ 更新代码中.. ]"):
                 move(temp_proj_p.joinpath(changed_file), existed_proj_p.joinpath(changed_file))
         shutil.rmtree(temp_p, onerror=delete)
+
+    def end(self):
         with open(existed_proj_p.joinpath('version'), 'w', encoding='utf-8') as f:
             f.write(self.ver)
+        print(Fore.GREEN + "=" * 40 + "[ 更新完毕 ]" + "=" * 40)
+
+
+def regular_update():
+    retry_times = 1
+    __ = None
+    while retry_times < 4:
+        try:
+            proj = Proj()
+            proj.check()
+            proj.local_update()
+            proj.end()
+            break
+        except Exception as e:
+            __ = traceback.format_exc()
+            print(Fore.RED + f"[ 更新失败, 准备重试-{retry_times} ]\n{type(e)} {e} ")
+            retry_times += 1
+    if retry_times > 3:
+        print(__)
+        print(Fore.RED + f"[Errno 11001] 是网络问题，重试更新即可。 若其他情况导致更新一直失败请截图发issue或找群反映")
 
 
 if __name__ == '__main__':
-    proj = Proj()
-    proj.check()
-    proj.local_update()
-    print(Fore.GREEN + "=" * 40 + "[ 更新完毕 ]" + "=" * 40)
+    regular_update()
